Moved test_utils files to test_utils directory and renamed them
