+++ targets/pna_nic/tests/test_parse_pna.cpp
@@ -22,7 +22,7 @@
 #include <bm/bm_apps/packet_pipe.h>
 #include <boost/filesystem.hpp>
 #include "pna_nic.h"
-#include "../../test_utils.h"
+#include "../../test_utils/utils.h"
 
 namespace fs = boost::filesystem;
