+++ targets/simple_switch/tests/test_packet_redirect.cpp
@@ -35,7 +35,7 @@
 
 #include "simple_switch.h"
 
-#include "../../test_utils.h"
+#include "../../test_utils/utils.h"
 
 namespace fs = boost::filesystem;
