+++ targets/simple_switch/tests/test_queueing.cpp
@@ -26,7 +26,7 @@
 
 #include "simple_switch.h"
 
-#include "../../test_utils.h"
+#include "../../test_utils/utils.h"
 
 namespace fs = boost::filesystem;
